split up styles

+++ server/modules/db.js
@@ -6,8 +6,6 @@ var dataPath = __dirname + '/../../content/data/data.json'
 
 var exists = fs.existsSync(dataPath)
 
-if(!exists) {
-	fs.writeFileSync(dataPath, '{"posts":[]}', 'utf8')
-}
+if(!exists) fs.writeFileSync(dataPath, '{"posts":[]}', 'utf8')
 
 module.exports = low(dataPath, { storage: storage })
